Refactor command package for external import and plugin support
