+++ cmd/alpamon/command/setup/configs/alpamon.conf
