+++ cmd/alpamon/command/setup/configs/tmpfile.conf
